Riders progress, links are now UUID, TMT changes
- Added a tool to check the latest version from git, it's pretty simple,
easily expandable, and works on any version of minecraft.
- Moved linking over to the new UUID system.
- Fixed a crash due to the concurrent checks on the entity list.
- More work trying to finish up the multiple passengers system.
- Fixed a packet not going to the server.
- Rider now gets scaled down slightly to better fit the
train/rollingstock.
- TMT change: fixed some invalid refrences, a couple typecasts were
actually necessary, oops.

+++ src/main/java/trains/models/tmt/ModelRendererTurbo.java
@@ -235,12 +235,12 @@ public class ModelRendererTurbo extends ModelRenderer {
         return this;
     }
     public ModelRendererTurbo addBox(float x, float y, float z, int w, int h, int d) {
-        addBox(x, y, z,  w,  h,  d, 0.0F);
+        addBox(x, y, z, (float) w, (float) h, (float) d, 0.0F);
         return this;
     }
     @Override
     public void addBox(float x, float y, float z, int w, int h, int d, float s) {
-        addBox(x, y, z,  w,  h,  d, s);
+        addBox(x, y, z, (float) w, (float) h, (float) d, s);
     }
